+++ src/Arius.Cli.Tests/DependencyInjectionTests.cs
@@ -3,7 +3,7 @@ using Arius.Core.Features.ArchiveCommand;
 using Arius.Core.Features.ChunkHydrationStatusQuery;
 using Arius.Core.Features.ContainerNamesQuery;
 using Arius.Core.Features.ListQuery;
-using Arius.Core.Features.Restore;
+using Arius.Core.Features.RestoreCommand;
 using Arius.Core.Shared.Storage;
 using Mediator;
 using Microsoft.Extensions.DependencyInjection;
