+++ src/Arius.Cli/Commands/Restore/RestoreProgressHandlers.cs
@@ -1,4 +1,4 @@
-using Arius.Core.Features.Restore;
+using Arius.Core.Features.RestoreCommand;
 using Mediator;
 
 namespace Arius.Cli.Commands.Restore;
