+++ src/Arius.Cli/Commands/Restore/RestoreVerb.cs
@@ -1,5 +1,5 @@
 using Arius.AzureBlob;
-using Arius.Core.Features.Restore;
+using Arius.Core.Features.RestoreCommand;
 using Arius.Core.Shared.Storage;
 using Humanizer;
 using Mediator;
