+++ src/Arius.Core.Tests/Restore/PricingConfigTests.cs
@@ -1,4 +1,4 @@
-using Arius.Core.Features.Restore;
+using Arius.Core.Features.RestoreCommand;
 using Shouldly;
 
 namespace Arius.Core.Tests.Restore;
