+++ src/Arius.Core/Arius.Core.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <EmbeddedResource Include="Features\Restore\pricing.json" />
+    <EmbeddedResource Include="Features\RestoreCommand\pricing.json" />
   </ItemGroup>
 
   <ItemGroup>
