+++ src/Arius.E2E.Tests/E2ETests.cs
@@ -1,6 +1,6 @@
 using Arius.AzureBlob;
 using Arius.Core.Features.ArchiveCommand;
-using Arius.Core.Features.Restore;
+using Arius.Core.Features.RestoreCommand;
 using Arius.Core.Shared.ChunkIndex;
 using Arius.Core.Shared.Encryption;
 using Arius.Core.Shared.Storage;
