+++ src/Arius.E2E.Tests/RehydrationE2ETests.cs
@@ -1,5 +1,5 @@
 using Arius.AzureBlob;
-using Arius.Core.Features.Restore;
+using Arius.Core.Features.RestoreCommand;
 using Arius.Core.Shared.ChunkIndex;
 using Arius.Core.Shared.Storage;
 using Shouldly;
