+++ src/Arius.Integration.Tests/Pipeline/RestoreDispositionTests.cs
@@ -1,4 +1,4 @@
-using Arius.Core.Features.Restore;
+using Arius.Core.Features.RestoreCommand;
 using Arius.Integration.Tests.Storage;
 using NSubstitute;
 using Shouldly;
