+++ src/Arius.Integration.Tests/Pipeline/RoundtripTests.cs
@@ -1,5 +1,5 @@
 using Arius.Core.Features.ArchiveCommand;
-using Arius.Core.Features.Restore;
+using Arius.Core.Features.RestoreCommand;
 using Arius.Core.Shared.Storage;
 using Arius.Integration.Tests.Storage;
 using Shouldly;
